+++ inst/tinytest/test-print.R
@@ -14,3 +14,15 @@ expect_snapshot_print(summary(mod1), "print-lm_summary")
 
 # Always fails
 expect_false(ignore(expect_snapshot_print)(summary(mod2), "print-lm_summary"))
+
+
+# Use fn_*() to remove random string from snapshot
+# Do NOT set a seed for this
+set.seed(NULL)
+x <- paste0("blah_", sample(1:1e6, 1))
+fn <- function(x) sub("_.*", "", x)
+expect_snapshot_print(x,
+    label = "print-sub_label_fn",
+    fn_current = fn,
+    fn_target = fn)
+